+++ cocoa-beans-commands/src/main/java/net/apartium/cocoabeans/commands/CommandBranchProcessor.java
@@ -121,11 +121,11 @@ import java.util.*;
         }
     }
 
-    /* package-private */ Set<String> handleTabCompletion(RegisteredCommand commandWrapper, String commandName, String[] args, Sender sender, int index) {
+    /* package-private */ Set<TabCompletionEvaluationResult> handleTabCompletion(RegisteredCommand commandWrapper, String commandName, String[] args, Sender sender, int index) {
         if (args.length <= index && index != 0)
             return Set.of();
 
-        Set<String> result = new HashSet<>();
+        Set<TabCompletionEvaluationResult> result = new HashSet<>();
         for (Entry<RequirementSet, CommandOption> entry : objectMap) {
             CommandOption commandOption = entry.value();
             if (commandOption == null)